[Js app structure] - Refactored folder names
