+++ client/js/app/main/common/header/index.js
