+++ client/js/app/main/common/menu/menu.template.html
