+++ client/js/app/main/modules/project/index.js
