+++ client/js/app/main/modules/project/project.config.js
