+++ client/js/app/main/modules/today/today.config.js
