+++ client/js/app/main/modules/today/today.view.html
