+++ client/js/app/main/modules/week/index.js
