+++ client/js/app/main/modules/week/week.config.js
